@@ -9,7 +9,8 @@
     "build": "ng build",
     "watch": "ng build --watch --configuration development",
     "test": "ng test",
-    "lint": "ng lint"
+    "lint": "ng lint",
+    "lint:fix": "ng lint --fix"
   },
   "private": true,
   "dependencies": {
@@ -33,6 +34,7 @@
     "@ionic/angular": "^8.0.0",
     "@ionic/pwa-elements": "^3.3.0",
     "@types/node": "^22.13.14",
+    "eslint-config-prettier": "^10.1.8",
     "ionicons": "^8.0.13",
     "marked": "^16.4.0",
     "ngx-markdown": "^20.1.0",
@@ -43,20 +45,16 @@
   },
   "devDependencies": {
     "@angular-devkit/build-angular": "^20.1.3",
-    "@angular-eslint/builder": "^20.1.1",
-    "@angular-eslint/eslint-plugin": "^20.1.1",
-    "@angular-eslint/eslint-plugin-template": "^20.1.1",
-    "@angular-eslint/schematics": "^20.1.1",
-    "@angular-eslint/template-parser": "^20.1.1",
+    "@angular-eslint/builder": "20.1.1",
     "@angular/cli": "^20.1.3",
     "@angular/compiler-cli": "^20.1.3",
     "@angular/language-service": "^20.1.3",
     "@capacitor/cli": "7.1.0",
+    "@eslint/js": "^9.29.0",
     "@ionic/angular-toolkit": "^12.0.0",
     "@types/jasmine": "~5.1.0",
-    "@typescript-eslint/eslint-plugin": "^8.18.0",
-    "@typescript-eslint/parser": "^8.18.0",
-    "eslint": "^9.16.0",
+    "angular-eslint": "20.1.1",
+    "eslint": "^9.29.0",
     "eslint-plugin-import": "^2.29.1",
     "eslint-plugin-jsdoc": "^48.2.1",
     "eslint-plugin-prefer-arrow": "1.2.2",
@@ -68,6 +66,7 @@
     "karma-jasmine": "~5.1.0",
     "karma-jasmine-html-reporter": "~2.1.0",
     "prettier": "^3.5.3",
-    "typescript": "~5.8.3"
+    "typescript": "~5.8.3",
+    "typescript-eslint": "8.34.1"
   }
 }
