@@ -1,14 +1,22 @@
-import { ErrorHandler, Injectable } from '@angular/core';
+import { ErrorHandler, inject, Injectable } from '@angular/core';
 import { ToastService } from '@services/toast.service';
 
 @Injectable({
   providedIn: 'root',
 })
 export class CbErrorHandler implements ErrorHandler {
-  constructor(private toastService: ToastService) {}
+  private readonly toastService = inject(ToastService);
 
+  // any is type specified by angular interface
+  // eslint-disable-next-line @typescript-eslint/no-explicit-any
   handleError(error: any): void {
     console.error('An error occurred:', error);
-    this.toastService.showToast(error.message);
+    let msg: string;
+    if (typeof error === 'string') {
+      msg = error;
+    } else {
+      msg = String(error);
+    }
+    this.toastService.showToast(msg);
   }
 }
