@@ -1,11 +1,11 @@
-import { Injectable } from '@angular/core';
+import { inject, Injectable } from '@angular/core';
 import { ToastController } from '@ionic/angular/standalone';
 
 @Injectable({
   providedIn: 'root',
 })
 export class ToastService {
-  constructor(private toastController: ToastController) {}
+  private readonly toastController = inject(ToastController);
 
   async showToast(message: string) {
     const toast = await this.toastController.create({
